BIC-methods {bbmle}	R Documentation

Log likelihoods and model selection for mle2 objects

Description

Various functions for likelihood-based and information-theoretic model selection of likelihood models

Usage



## S4 method for signature 'ANY,mle2,logLik'
AICc(object,...,nobs,k=2)
## S4 method for signature 'ANY,mle2,logLik'
qAIC(object,...,k=2)
## S4 method for signature 'ANY,mle2,logLik'
qAICc(object,...,nobs,k=2)

Arguments

object
A logLik or mle2 object

...
An optional list of additional logLik or mle2 objects (fitted to the same data set).

nobs
Number of observations (sometimes obtainable as an attribute of the fit or of the log-likelihood)

k
penalty parameter (nearly always left at its default value of 2)

Details

Further arguments to BIC can be specified in the ... list: delta (logical) specifies whether to include a column for delta-BIC in the output.

Value

A table of the BIC values, degrees of freedom, and possibly delta-BIC values relative to the minimum-BIC model

Methods

logLik

signature(object = "mle2"): Extract maximized log-likelihood.

AIC

signature(object = "mle2"): Calculate Akaike Information Criterion

AICc

signature(object = "mle2"): Calculate small-sample corrected Akaike Information Criterion

anova

signature(object="mle2"): Likelihood Ratio Test comparision of different models

Note

This is implemented in an ugly way and could probably be improved!

Examples


  d <- data.frame(x=0:10,y=c(26, 17, 13, 12, 20, 5, 9, 8, 5, 4, 8))
  (fit <- mle2(y~dpois(lambda=ymax/(1+x/xhalf)),
      start=list(ymax=25,xhalf=3),data=d))
  (fit2 <- mle2(y~dpois(lambda=(x+1)*slope),
      start=list(slope=1),data=d))
  BIC(fit)
  BIC(fit,fit2)
  

[Package bbmle version 1.0.24 Index]