dnorm_n {bbmle}	R Documentation

Normal distribution with profiled-out standard deviation

Description

Returns the Normal probability densities for a distribution with the given mean values and the standard deviation equal to the root mean-squared deviation between x and mu

Usage


dnorm_n(x, mean, log = FALSE)

Arguments

x
numeric vector of data

mean
numeric vector or mean values

log
logical: return the log-density?

Details

This is a convenience function, designed for the case where you're trying to compute a MLE for the mean but don't want to bother estimating the MLE for the standard deviation at the same time

Value

Numeric vector of probability densities

Examples


set.seed(101)
x <- rnorm(5,mean=3,sd=2)
dnorm_n(x,mean=3,log=TRUE)

[Package bbmle version 1.0.24 Index]