mle2-class {bbmle}	R Documentation

Class "mle2". Result of Maximum Likelihood Estimation.

Description

This class encapsulates results of a generic maximum likelihood procedure.

Objects from the Class

Objects can be created by calls of the form new("mle2", ...), but most often as the result of a call to mle2.

Slots

call:

(language) The call to mle2.

call.orig:

(language) The call to mle2, saved in its original form (i.e. without data arguments evaluated).

coef:

(numeric) Vector of estimated parameters.

data:

(data frame or list) Data with which to evaluate the negative log-likelihood function

fullcoef:

(numeric) Fixed and estimated parameters.

vcov:

(numeric matrix) Approximate variance-covariance matrix, based on the second derivative matrix at the MLE.

min:

(numeric) Minimum value of objective function = minimum negative log-likelihood.

details:

(list) Return value from optim.

minuslogl:

(function) The negative log-likelihood function.

optimizer:

(character) The optimizing function used.

method:

(character) The optimization method used.

formula:

(character) If a formula was specified, a character vector giving the formula and parameter specifications.

Methods

coef

signature(object = "mle2"): Extract coefficients. If exclude.fixed=TRUE (it is FALSE by default), only the non-fixed parameter values are returned.

confint

signature(object = "mle2"): Confidence intervals from likelihood profiles, or quadratic approximations, or root-finding.

show

signature(object = "mle2"): Display object briefly.

show

signature(object = "summary.mle2"): Display object briefly.

summary

signature(object = "mle2"): Generate object summary.

update

signature(object = "mle2"): Update fit.

vcov

signature(object = "mle2"): Extract variance-covariance matrix.

formula

signature(object="mle2"): Extract formula

plot

signature(object="profile.mle2,missing"): Plot profile.

Details on the confint method

When the parameters in the original fit are constrained using lower or upper, or when prof.lower or prof.upper are set, and the confidence intervals lie outside the constraint region, confint will return NA. This may be too conservative – in some cases, the appropriate answer would be to set the confidence limit to the lower/upper bound as appropriate – but it is the most general answer.

(If you have a strong opinion about the need for a new option to confint that sets the bounds to the limits automatically, please contact the package maintainer.)

Examples


x <- 0:10
y <- c(26, 17, 13, 12, 20, 5, 9, 8, 5, 4, 8)
lowerbound <- c(a=2,b=-0.2)
d <- data.frame(x,y)
fit1 <- mle2(y~dpois(lambda=exp(a+b*x)),start=list(a=0,b=2),data=d,
method="L-BFGS-B",lower=c(a=2,b=-0.2))
(cc <- confint(fit1,quietly=TRUE))
## to set the lower bounds to the limit
na_lower <- is.na(cc[,1])
cc[na_lower,1] <- lowerbound[na_lower]
cc

[Package bbmle version 1.0.24 Index]