mle2.options {bbmle}	R Documentation

Options for maximum likelihood estimation

Description

Query or set MLE parameters

Usage


mle2.options(...)

Arguments

...
names of arguments to query, or a list of values to set

Details

optim.method

name of optimization method (see optim for choices)

confint

name of confidence interval method: choices are "spline", "uniroot", "hessian" corresponding to spline inversion, attempt to find best answer via uniroot, information-matrix approximation

optimizer

optimization function to use by default (choices: "optim", "nlm", "nlminb", "constrOptim")

Value

Values of queried parameters, or (invisibly) the full list of parameters

See Also

mle2-class

[Package bbmle version 1.0.24 Index]