profile-methods {bbmle}	R Documentation

Likelihood profiles

Description

Compute likelihood profiles for a fitted model

Usage


proffun(fitted, which = 1:p, maxsteps = 100,
                    alpha = 0.01, zmax = sqrt(qchisq(1 - alpha/2, p)),
                    del = zmax/5, trace = FALSE, skiperrs=TRUE,
                    std.err, 
                    tol.newmin = 0.001, debug=FALSE,
                    prof.lower, prof.upper,
                    skip.hessian = TRUE,
                    continuation = c("none","naive","linear"),
                    try_harder=FALSE, ...)
## S4 method for signature 'mle2'
profile(fitted, ...)

Arguments

fitted
A fitted maximum likelihood model of class “mle2”

which
a numeric or character vector describing which parameters to profile (default is to profile all parameters)

maxsteps
maximum number of steps to take looking for an upper value of the negative log-likelihood

alpha
maximum (two-sided) likelihood ratio test confidence level to find

zmax
maximum value of signed square root of deviance difference to find (default value corresponds to a 2-tailed chi-squared test at level alpha)

del
step size for profiling

trace
(logical) produce tracing output?

skiperrs
(logical) ignore errors produced during profiling?

std.err
Optional numeric vector of standard errors, for cases when the Hessian is badly behaved. Will be replicated if necessary, and NA values will be replaced by the corresponding values from the fit summary

tol.newmin
tolerance for diagnosing a new minimum below the minimum deviance estimated in initial fit is found

debug
(logical) debugging output?

prof.lower
optional vector of lower bounds for profiles

prof.upper
optional vector of upper bounds for profiles

continuation
use continuation method to set starting values? "none" sets starting values to best fit; "naive" sets starting values to those of previous profiling fit; "linear" (not yet implemented) would use linear extrapolation from the previous two profiling fits

skip.hessian
skip hessian (defunct?)

try_harder
(logical) ignore NA and flat spots in the profile, try to continue anyway?

...
additional arguments (not used)

Details

proffun is the guts of the profile method, exposed so that other packages can use it directly.

See the vignette (vignette("mle2",package="bbmle")) for more technical details of how profiling is done.

See Also

profile.mle-class

[Package bbmle version 1.0.24 Index]