sbinom {bbmle}	R Documentation

Abstract definitions of distributions

Description

Functions returning values for summary statistics (mean, median, etc.) of distributions

Usage


sbeta(shape1, shape2)
sbetabinom(size, prob, theta)
sbinom(size, prob)
snbinom(size, prob, mu)
snorm(mean, sd)
spois(lambda)
slnorm(meanlog, sdlog)

Arguments

prob
probability as defined for dbinom, dnbinom, or beta-binomial distribution (dbetabinom in the emdbook package)

size
size parameter as defined for dbinom or dbetabinom in the emdbook package, or size/overdispersion parameter as in dnbinom

mean
mean parameter as defined for dnorm

mu
mean parameter as defined for dnbinom

sd
standard deviation parameter as defined for dnorm

shape1
shape parameter for dbeta

shape2
shape parameter for dbeta

lambda
rate parameter as defined for dpois

theta
overdispersion parameter for beta-binomial (see dbetabinom in the emdbook package)

meanlog
as defined for dlnorm

sdlog
as defined for dlnorm

Value

title
name of the distribution

[parameters]
input parameters for the distribution

mean
theoretical mean of the distribution

median
theoretical median of the distribution

mode
theoretical mode of the distribution

variance
theoretical variance of the distribution

sd
theoretical standard deviation of the distribution

Note

these definitions are tentative, subject to change as I figure this out better. Perhaps construct functions that return functions? Strip down results? Do more automatically?

Author(s)

Ben Bolker

See Also

dbinom, dpois, dnorm, dnbinom

Examples


  sbinom(prob=0.2,size=10)
  snbinom(mu=2,size=1.2)

[Package bbmle version 1.0.24 Index]