slice.mle2-class {bbmle}	R Documentation

likelihood-surface slices

Description

evaluations of log-likelihood along transects in parameter space

Objects from the Class

Objects can be created by calls of the form new("slice.mle2", ...). The objects are similar to likelihood profiles, but don't involve any optimization with respect to the other parameters.

Slots

profile:

Object of class "list". List of slices, one for each requested parameter. Each slice is a data frame with the first column called z being the signed square root of the -2 log likelihood ratio, and the others being the parameters with names prefixed by par.vals.

summary:

Object of class "summary.mle2". Summary of object being profiled.

Methods

plot

signature(x = "profile.mle2", y = "missing"): Plot profiles for each parameter.

See Also

profile.mle2-class

[Package bbmle version 1.0.24 Index]